Improved communities and metrics performance

+++ scaner/compute_communities.py
@@ -24,44 +24,46 @@ def execution():
 	
 	for r in client.query('select from follows limit -1'):
 		edges.append((str(r.oRecordData['out']), str(r.oRecordData['in'])))
-
-	G = nx.Graph()
-	G.add_nodes_from(nodes)
-	G.add_edges_from(edges)
-	nodes = G.nodes()
-
-
-	import community
-
-	partition = community.best_partition(G)
-	#partition2 = community.generate_dendrogram(G, part_init=None, weight='weight', resolution=1.0)
-	#drawing
-	#size = float(len(set(partition.values())))
-	#pos = nx.spring_layout(G)
-	#count = 0.
-	#for com in set(partition.values()) :
-	#    count += 1.
-	#    list_nodes = [nodes for nodes in partition.keys()
-	#                                if partition[nodes] == com]
-	#    nx.draw_networkx_nodes(G, pos, list_nodes, node_size = 20,
-	#                            node_color = str(count / size))
-	#nx.draw_networkx_edges(G,pos, alpha=0.5)
-
-	client.command("delete edge Belongs_to_Community from (select from User) to (select from Community)")
-	client.command("delete from Community UNSAFE")
-	reg = []
-	for item in partition:
-	    if partition[item] not in reg:
-	        client.command("insert into Community set id= {id}, user_count=0".format(id=partition[item]))
-	    client.command("update user set community = {community} where @rid = '{rid}'".format(community=partition[item], rid=item))
-	    client.command("create edge Belongs_to_Community from (select from User where @rid = {rid}) to (select from Community where id = {id})".format(rid=item, id =partition[item]))
-	    reg.append(partition[item])
+	if edges and nodes:
+		G = nx.Graph()
+		G.add_nodes_from(nodes)
+		G.add_edges_from(edges)
+		nodes = G.nodes()
 
 
-	communities = client.query("select from community limit -1")
-	for community in communities:
-	    count = client.query("select count(*) from user where community = {id}".format(id=community.oRecordData['id']))
-	    client.command("update Community set user_count = {user_count} where id = {id}".format(user_count = count[0].oRecordData['count'], id =community.oRecordData['id']))
+		import community
+
+		partition = community.best_partition(G)
+		#partition2 = community.generate_dendrogram(G, part_init=None, weight='weight', resolution=1.0)
+		#drawing
+		#size = float(len(set(partition.values())))
+		#pos = nx.spring_layout(G)
+		#count = 0.
+		#for com in set(partition.values()) :
+		#    count += 1.
+		#    list_nodes = [nodes for nodes in partition.keys()
+		#                                if partition[nodes] == com]
+		#    nx.draw_networkx_nodes(G, pos, list_nodes, node_size = 20,
+		#                            node_color = str(count / size))
+		#nx.draw_networkx_edges(G,pos, alpha=0.5)
+
+		client.command("delete edge Belongs_to_Community from (select from User) to (select from Community)")
+		client.command("delete from Community UNSAFE")
+		reg = []
+		for item in partition:
+		    if partition[item] not in reg:
+		        client.command("insert into Community set id= {id}, user_count=0".format(id=partition[item]))
+		    client.command("update user set community = {community} where @rid = '{rid}'".format(community=partition[item], rid=item))
+		    client.command("create edge Belongs_to_Community from (select from User where @rid = {rid}) to (select from Community where id = {id})".format(rid=item, id =partition[item]))
+		    reg.append(partition[item])
 
 
-	print("FINISH :::::TASK::::::")
+		communities = client.query("select from community limit -1")
+		for community in communities:
+		    count = client.query("select count(*) from user where community = {id}".format(id=community.oRecordData['id']))
+		    client.command("update Community set user_count = {user_count} where id = {id}".format(user_count = count[0].oRecordData['count'], id =community.oRecordData['id']))
+
+
+		print("FINISH :::::TASK::::::")
+	else:
+		print("Error: You do not have a network between users in yout database")
